✦ Minor UI Improvement (#114)
* Change Date Format in Charts #106

* Fix DepositForm Loading #102

DepositForm now will only go to loading state when USDC balance is loading. So when the exchange rate is still loading, it will load the DepositForm and display a spinner in exchange rate value instead.

* Add Contract Address & Link to Explorer for Leveraged Token #104

* Rename 'Earn' to 'Lend' in ETHRISE Page

* Run Prettier Write

* Change Date Format in Market Page Chart

* Fix 'href' in Leveraged Token Contract Information

* Fix 'href' in Leveraged Token Contract Information 2

+++ components/v1/MarketsPage/MarketCard.tsx
@@ -106,7 +106,7 @@ const MarketCard: FunctionComponent<MarketCardProps> = ({ address, initialNAV, i
                                         const selectedData = payload[0].payload;
                                         const timestamp = selectedData.timestamp;
                                         const date = new Date(timestamp);
-                                        const formattedDate = new Intl.DateTimeFormat("en-US", { hour: "numeric", day: "numeric", month: "numeric", year: "numeric", minute: "numeric" }).format(date);
+                                        const formattedDate = new Intl.DateTimeFormat("en-US", { hour: "numeric", day: "numeric", month: "short", year: "numeric", minute: "numeric" }).format(date);
 
                                         setNAV(selectedData.nav);
                                         const change = ((selectedData.nav - currentData.oldestNAV) / currentData.oldestNAV) * 100;
